Create new plugin to load data
