+++ src/plugins/unpkgPathPlugin.ts
@@ -1,12 +1,6 @@
 import * as esbuild from 'esbuild-wasm';
-import axios from 'axios';
-import localforage from 'localforage'
 
-const fileCache = localforage.createInstance({
-   name: 'filecache'
-})
-
-export const unpkgPathPlugin = (input: string) => {
+export const unpkgPathPlugin = () => {
    return {
       name: "unpkg-path-plugin",
       setup(build: esbuild.PluginBuild) {
@@ -30,35 +24,6 @@ export const unpkgPathPlugin = (input: string) => {
                namespace: 'a'
             }
          })
-
-         build.onLoad({ filter: /.*/ }, async (args: any) => {
-            console.log('on Load', args)
-
-            if (args.path === 'index.js') {
-               return {
-                  loader: 'jsx',
-                  contents: input,
-               };
-            }
-
-            const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(args.path)
-
-            if (cachedResult) {
-               return cachedResult;
-            }
-
-            const { data, request } = await axios.get(args.path);
-
-            const result: esbuild.OnLoadResult = {
-               loader: 'jsx',
-               contents: data,
-               resolveDir: new URL('./', request.responseURL).pathname
-            }
-
-            await fileCache.setItem(args.path, result);
-
-            return result;
-         });
       },
    };
 };
